+++ crates/sdk/src/std/network-v2/proto/artifact.rs
