+++ crates/sdk/src/std/network/prover.rs
@@ -1,5 +1,6 @@
 use std::{
     env,
+    string::String,
     time::{Duration, Instant},
 };
